[Refactor] Rename service packages to improve organization and clarity
