+++ apigateway/src/main/java/vaultweb/apigateway/controller/GatewayAuthController.java
@@ -15,7 +15,7 @@ import vaultweb.apigateway.dto.request.UserRegistrationRequest;
 import vaultweb.apigateway.dto.response.AuthResponse;
 import vaultweb.apigateway.dto.response.UserDetails;
 import vaultweb.apigateway.exceptions.DefaultException;
-import vaultweb.apigateway.service.AuthService;
+import vaultweb.apigateway.service.auth.AuthService;
 
 @RestController
 @RequestMapping("/auth")
